+++ NEWS.md
@@ -1,3 +1,9 @@
+# ppgmmga 1.0.2 (2018-12) NOT ON CRAN
+
+- Fix a warning in `plot.ppgmmga()` when d > 2 and `dim` is set to two 
+  coordinates.
+- Update vignette.
+  
 # ppgmmga 1.0.1 (2018-10)
 
 - Fix a C++ issue.